Move password utilities to their own file
