+++ webapp/channels/src/components/admin_console/reset_password_modal/index.ts
@@ -6,21 +6,17 @@ import {bindActionCreators} from 'redux';
 import type {Dispatch} from 'redux';
 
 import {updateUserPassword} from 'mattermost-redux/actions/users';
-import {getConfig} from 'mattermost-redux/selectors/entities/general';
+import {getPasswordConfig} from 'mattermost-redux/selectors/entities/general';
 import {getCurrentUserId} from 'mattermost-redux/selectors/entities/users';
-
-import {getPasswordConfig} from 'utils/utils';
 
 import type {GlobalState} from 'types/store';
 
 import ResetPasswordModal from './reset_password_modal';
 
 function mapStateToProps(state: GlobalState) {
-    const config = getConfig(state);
-
     return {
         currentUserId: getCurrentUserId(state),
-        passwordConfig: getPasswordConfig(config),
+        passwordConfig: getPasswordConfig(state),
     };
 }
