+++ webapp/channels/src/components/user_settings/security/user_settings_security.test.tsx
@@ -7,9 +7,10 @@ import React from 'react';
 import type {OAuthApp} from '@mattermost/types/integrations';
 import type {UserProfile} from '@mattermost/types/users';
 
+import type {PasswordConfig} from 'mattermost-redux/selectors/entities/general';
+
 import type {MockIntl} from 'tests/helpers/intl-test-helper';
 import Constants from 'utils/constants';
-import type * as Utils from 'utils/utils';
 
 import {SecurityTab} from './user_settings_security';
 
@@ -46,7 +47,7 @@ describe('components/user_settings/display/UserSettingsDisplay', () => {
         enableSaml: true,
         enableSignUpWithOffice365: false,
         experimentalEnableAuthenticationTransfer: true,
-        passwordConfig: {} as ReturnType<typeof Utils.getPasswordConfig>,
+        passwordConfig: {} as PasswordConfig,
         militaryTime: false,
         intl: {
             formatMessage: jest.fn(({id, defaultMessage}) => defaultMessage || id),
